Support input files that are not in the current directory
Depending on the exact OpenSCAD version used, different behavior resulted if the input file was not in the current directory.
Sometimes "use" and "import" statements didn't work, other times the intermediate ".csg" output was saved at an unexpected place.
Work around it, by changing the current directory to the one containing the input file.